+++ app/src/main/java/ru/netology/nmedia/repository/PostRepositoryInMemoryImpl.kt
@@ -1,0 +1,34 @@
+package ru.netology.nmedia.repository
+
+import ru.netology.nmedia.dto.Post
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+
+
+class PostRepositoryInMemoryImpl : PostRepository {
+
+    private var post = Post(
+        id = 1L,
+        author = "Нетология. Университет интернет-профессий будущего",
+        content = "Привет, это новая Нетология! Когда-то Нетология начиналась с интенсивов по онлайн-маркетингу. Затем появились курсы по дизайну, разработке, аналитике и управлению. Мы растём сами и помогаем расти студентам: от новичков до уверенных профессионалов. Но самое важное остаётся с нами: мы верим, что в каждом уже есть сила, которая заставляет хотеть больше, целиться выше, бежать быстрее. Наша миссия — помочь встать на путь роста и начать цепочку перемен → http://netolo.gy/fyb",
+        published = "21 мая в 18:36",
+        likedByMe = false,
+        countLikes = 999u,
+        countShare = 1999u,
+        countGlaz = 999999u
+    )
+
+    private val data = MutableLiveData(post)
+
+    override fun get(): LiveData<Post> = data
+
+    override fun like() {
+        post = post.copy(likedByMe = !post.likedByMe)
+        if (post.likedByMe) {
+            post.countLikes++
+        } else {
+            post.countLikes--
+        }
+        data.value = post
+    }
+}
